post creation fixes

+++ db/post.go
@@ -6,7 +6,7 @@ type Post struct {
 	ID        bson.ObjectId `bson:"_id" json:"key"`
 	Title     string        `bson:"title" json:"title"`
 	Body      string        `bson:"body" json:"body"`
-	Timestamp int           `bson:"timestamp" json:"timestamp"`
+	Timestamp int64         `bson:"timestamp" json:"timestamp"`
 	Frontpage bool          `bson:"frontpage" json:"frontpage"`
 	UserID    string        `bson:"user_id" json:"user_id"`
 }
